+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Changelog
 Development
 -----------
 
+* Change .iteritems() to .items() in accordance with pandas>=2.0.0
+* .get_loc(x, method=...) to .get_indexer([x],method=...)[0] in accordance with pandas>=2.0.0
+* Updated mean() to mean(numeric_only=True) in accordance to pandas>=2.0.0
+* Updated tests to work with pandas>=2.0.0
 * Update python version in Dockerfile.
 * Update other dependencies (including adding rust) in Dockerfile.
 * Remove pinned dependencies in Pipfile.